requirements.txt
discord.py 
pyyaml
nltk
torch
numpy
pyllamacpp